+++ mediaserver/src/motion/motion_archive.cpp
@@ -152,7 +152,7 @@ void QnMotionArchive::maskMotion(QnMetaDataV1Ptr data)
 void QnMotionArchive::updateMotionMask(QRegion maskedRegion)
 {
     QMutexLocker lock(&m_maskMutex);
-    Q_ASSERT(((unsigned int)m_motionMask)%16 == 0);
+    Q_ASSERT(((unsigned long)m_motionMask)%16 == 0);
     QnMetaDataV1::createMask(maskedRegion, (char*)m_motionMask, &m_motionMaskStart, &m_motionMaskEnd);
 }
 
@@ -204,7 +204,7 @@ QnTimePeriodList QnMotionArchive::mathPeriod(const QRegion& region, qint64 msSta
     __m128i mask[MD_WIDTH * MD_HEIGHT / 128];
     int maskStart, maskEnd;
 
-    Q_ASSERT(((unsigned int)mask)%16 == 0);
+    Q_ASSERT(((unsigned long)mask)%16 == 0);
 
     QnMetaDataV1::createMask(region, (char*)mask, &maskStart, &maskEnd);
     bool isFirstStep = true;
